refactor(background): rename emitter observer spelling error
