+++ src/background/emitter_observer.js
