+++ src/background/main.js
@@ -1,7 +1,7 @@
 import timer from './timer'
 import { pomodoro_store, events } from './pomodoro_store'
 import { settings_writable, settings_readable } from './settings_store'
-import emitter_observer from './emmitter_observer'
+import emitter_observer from './emitter_observer'
 // import history_observer from './history_observer'
 // import notification_observer from './notification_observer'
 // import timer_sound_observer from './timer_sound_observer'
